Core

Core functionality for PhysiCellModelManager.jl.

Public API

ModelManager.initializeModelManager — Method

initializeModelManager()
initializeModelManager(path_to_project_dir::AbstractString)
initializeModelManager(path_to_physicell::AbstractString, path_to_data::AbstractString)

Initialize the PhysiCellModelManager.jl project model manager, identifying the data folder, PhysiCell folder, and loading the central database.

If no arguments are provided, it assumes that the PhysiCell and data directories are inside the current working directory.

Arguments

path_to_project_dir::AbstractString: Path to the project directory. Must contain the PhysiCell and data subdirectories.

path_to_physicell::AbstractString: Path to the PhysiCell directory.

path_to_data::AbstractString: Path to the data directory.

Private API

PhysiCellModelManager.PCMMMissingProject — Type

PCMMMissingProject

An exception type for when a PhysiCellModelManager.jl project cannot be found during initialization.

Fields

msg::String: The error message.

PhysiCellModelManager.pcmmLogo — Method

pcmmLogo()

Return a string representation of the awesome PhysiCellModelManager.jl logo.