Movie

Make movies for simulations in the database.

Public API

PhysiCellModelManager.makeMovie — Method

makeMovie(simulation_id::Int; magick_path::Union{Missing,String}=pcmm_globals.path_to_magick, ffmpeg_path::Union{Missing,String}=pcmm_globals.path_to_ffmpeg)
makeMovie(T::AbstractTrial; kwargs...)
makeMovie(out::PCMMOutput; kwargs...)

Batch make movies for each simulation identified by the input.

Use the PhysiCell Makefile to generate the movie. This process requires first generating JPEG files, which are then used to create the movie. Deletes the JPEG files after the movie is generated.

This relies on ImageMagick and FFmpeg being installed on the system. There are three ways to allow this function to find these dependencies:

Pass the path to the dependencies using the magick_path and ffmpeg_path keyword arguments.

Set the PATH environment variable to include the directories containing the dependencies.

Set environment variables PCMM_IMAGEMAGICK_PATH and PCMM_FFMPEG_PATH before using PhysiCellModelManager.

Arguments

simulation_id::Int: The ID of the simulation for which to make the movie.

T::AbstractTrial: Make movies for all simulations in the AbstractTrial.

out::PCMMOutput: Make movies for all simulations in the output, i.e., all simulations in the completed trial.

Keyword Arguments

magick_path::Union{Missing,String}: The path to the ImageMagick executable. If not provided, uses the global variable pcmm_globals.path_to_magick.

ffmpeg_path::Union{Missing,String}: The path to the FFmpeg executable. If not provided, uses the global variable pcmm_globals.path_to_ffmpeg.

Example

makeMovie(123) # make a movie for simulation 123

makeMovie(sampling) # make movies for all simulations in the sampling

out = run(sampling) # run the sampling
makeMovie(out) # make movies for all simulations in the output

Private API

PhysiCellModelManager.resolveMovieGlobals — Method

resolveMovieGlobals(magick_path::Union{Missing,String}, ffmpeg_path::Union{Missing,String})

Set the global variables path_to_magick and path_to_ffmpeg to the provided paths.