PhysiCellModelManager.jl Version

Manage the version of PhysiCellModelManager.jl in the database.

Public API

Private API

PhysiCellModelManager.pcmmDBVersion — Method

pcmmDBVersion()

Returns the version of the PhysiCellModelManager.jl database. If the database does not exist, it creates a new one with the current PhysiCellModelManager.jl version.

PhysiCellModelManager.pcmmVersion — Method

pcmmVersion()

Returns the version of the PhysiCellModelManager.jl package.

PhysiCellModelManager.resolvePCMMVersion — Method

resolvePCMMVersion(auto_upgrade::Bool)

Resolve differences between the PhysiCellModelManager.jl version and the database version. If the PhysiCellModelManager.jl version is lower than the database version, it returns false (upgrade your version of PhysiCellModelManager.jl to match what was already used for the database). If the PhysiCellModelManager.jl version is equal to the database version, it returns true. If the PhysiCellModelManager.jl version is higher than the database version, it upgrades the database to the current PhysiCellModelManager.jl version and returns true.

PhysiCellModelManager.versionFromTable — Method

versionFromTable(table_name::String; kwargs...)

Returns the version from the specified table if it exists, otherwise returns nothing. The kwargs... are passed to the tableExists function to check if the table exists.